Allow client code to use protocols on their RLMObject properties.
Allow client code to have properties which are declared with protocols
other than the Realm ones (RLMInt etc).  This means that client code
can use protocols to indicate other attributes of their properties.

Adopt the convention that the RLM* protocol must be last in the list
of protocols, to make parsing easier and faster inside the reflection
code.